+++ berkshelf/berkshelf.go
@@ -5,15 +5,15 @@ import (
 	"github.com/mlafeldt/chef-runner/util"
 )
 
-func InstallCommand(path string) []string {
+func Command(dst string) []string {
 	var cmd []string
 	if util.FileExist("Gemfile") {
 		cmd = []string{"bundle", "exec"}
 	}
-	cmd = append(cmd, "berks", "install", "--path", path)
+	cmd = append(cmd, "berks", "install", "--path", dst)
 	return cmd
 }
 
-func Install(path string) error {
-	return exec.RunCommand(InstallCommand(path))
+func InstallCookbooks(dst string) error {
+	return exec.RunCommand(Command(dst))
 }