+++ berkshelf/berkshelf_test.go
@@ -8,18 +8,18 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestInstallCommand(t *testing.T) {
+func TestCommand(t *testing.T) {
 	expect := []string{"berks", "install", "--path", "a/b/c"}
-	actual := berkshelf.InstallCommand("a/b/c")
+	actual := berkshelf.Command("a/b/c")
 	assert.Equal(t, expect, actual)
 }
 
-func TestInstallCommand_Bundler(t *testing.T) {
+func TestCommand_Bundler(t *testing.T) {
 	f, _ := os.Create("Gemfile")
 	f.Close()
 	defer os.Remove("Gemfile")
 
 	expect := []string{"bundle", "exec", "berks", "install", "--path", "a/b/c"}
-	actual := berkshelf.InstallCommand("a/b/c")
+	actual := berkshelf.Command("a/b/c")
 	assert.Equal(t, expect, actual)
 }